# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Amana Backend API
  version: 1.0.0
  description: |-
    OpenAPI coverage for the live Amana backend routes under `backend/src/routes`.

    Authentication flow:
    1. Call `POST /auth/challenge` with a Stellar public key.
    2. Sign the returned challenge with the wallet.
    3. Exchange the signature at `POST /auth/verify` for a bearer token.
    4. Send `Authorization: Bearer <token>` to protected endpoints.

    JWT semantics enforced by middleware:
    - `iss` must match `JWT_ISSUER` (default `amana`)
    - `aud` must match `JWT_AUDIENCE` (default `amana-api`)
    - `jti` is required and checked against the revocation denylist
      (revoked tokens are rejected — replay protection)
    - `iat` (issued-at) must be present and in the past
    - `exp` (expires-at) is enforced: expired tokens are rejected
      (default TTL: 86400 s / 24 h, configurable via `JWT_EXPIRES_IN`)
    - `nbf` (not-before) must be present and not be in the future
    - Tokens signed with algorithms other than HS256 are rejected

    Error code references in operation descriptions map to backend `ErrorCode` values in
    `backend/src/errors/errorCodes.ts`. Some route handlers still emit legacy `{ error }`
    payloads while centralized validation/runtime failures return `{ code, message, details }`.

    # API Versioning

    The public API is served on two lanes that share identical behaviour:

...
      e.g. `POST /api/v1/auth/login`.
...
      not yet moved to `/api/v1`.

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Secret key authentication for admin-only management routes.
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example:
        error: Unauthorized
...
  responses:
    UnauthorizedError:
      description: Missing, invalid, revoked, or semantically invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unauthorized:
              value:
                error: Unauthorized
            revoked:
              value:
                error: "Unauthorized: token has been revoked"
    ForbiddenError:
      description: Authenticated caller is not allowed to perform the action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error: Forbidden
...
paths:
...
  /api/admin/auth/claims:
    get:
      tags: [Admin]
      summary: Inspect the caller's parsed admin JWT claims
      description: |-
        Diagnostic endpoint returning the sanitized JWT claims the backend
        parsed from the caller's bearer token (wallet address, token id,
        issuer/audience, issued/expiry timestamps), so an admin can verify
        exactly what the server sees without decoding the token by hand.
        Raw fields not meaningful to display (`sub`, `nbf`) are omitted.
        Error code references: `AUTH_ERROR`.
      security:
        - bearerAuth: []
        - adminAuth: []
      responses:
        "200":
          description: Sanitized admin JWT claims.
          content:
            application/json:
              schema:
                type: object
                properties:
                  walletAddress:
                    type: string
                  tokenId:
                    type: string
                  issuedAt:
                    type: string
                    nullable: true
                  expiresAt:
                    type: string
                    nullable: true
                  issuer:
                    type: string
                    nullable: true
                  audience:
                    type: string
                    nullable: true
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /api/admin/sessions/revoke:
    post:
      tags: [Admin]
      summary: Revoke an active admin session or token JTI
      description: |-
        Control plane endpoint to revoke active admin session JTIs.
      security:
        - bearerAuth: []
        - adminAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                jti:
                  type: string
                  description: Optional session JTI to invalidate (defaults to caller's JTI).
      responses:
        "200":
          description: Admin session revoked successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  jti:
                    type: string
                  revokedAt:
                    type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...